BXMSPROD-894 main.js refactoring
https://issues.redhat.com/browse/BXMSPROD-894
