+++ src/lib/pr.js
@@ -1,0 +1,19 @@
+const globToRegExp = require('glob-to-regexp');
+const { getChangedFiles } = require("./utils");
+
+async function isFirstPR(context) {
+    const respone = await context.github.issues.listForRepo(context.repo({
+        state: 'all',
+        creator: context.payload.pull_request.user.login
+    }))
+    const countPR = respone.data.filter(data => data.pull_request);
+    if (countPR.length === 1) {
+        return true
+    } else {
+        return false
+    }
+}
+
+module.exports = {
+    isFirstPR
+};
